refactor(ir): rename IrExtension.transform to extend
Address review feedback: rename the IrExtension method `transform` to
`extend` so it reads as extending the IR rather than the internal AST
transform machinery. Update the interface, the IrEmitter fold call, and
the Spring + kotlinx-serialization implementations.

Also move ExtendingIrEmitterTest into the extension test package, which
the earlier package move missed (it resolved the symbols same-package).
